Write colour in the linear space glTF asks for
glTF defines COLOR_0 as linear and scanner colour is sRGB encoded, so passing
the source values through unchanged made every tile wash out. The pipeline this
replaces was converting on the way out of Blender, which is why the difference
only showed up against an older tileset in the same scene.

Colours are 16 bit now. Eight bits of linear bands visibly in the dark half,
which on a scan is most of it, and 16 bit is what the old path emitted anyway.
Costs four bytes per point where a scan has colour.

Checked against a tileset built by the old pipeline from the same scan: 38,242
points, every one within 0.00002 of the old value, which is the rounding.

--rawColor opts out for a source whose colour is already linear.

+++ entwine/formats/cesium/gltf.hpp
@@ -32,7 +32,10 @@ namespace cesium
 class Gltf
 {
     using Xyz = std::vector<float>;
-    using Rgba = std::vector<uint8_t>;
+
+    // 16 bit, because these are linear values: 8 bits of linear would band
+    // badly in the dark half, where most of a scan lives.
+    using Rgba = std::vector<uint16_t>;
 
 public:
     Gltf(const Tileset& tileset, const ChunkKey& ck, uint64_t np);
